+++ spec/app/executions_spec.rb
@@ -315,5 +315,46 @@ describe '/executions' do
       end
     end
   end
+
+  describe 'DELETE /executions/:exid' do
+
+    before :each do
+
+      @exids = %w[ net.ntt net.ntt.hr ]
+        .collect { |d| @app.unit.launch(%{ sleep '1d' }, domain: d) }
+        .sort
+      @app.unit.wait('idle')
+    end
+
+    it 'goes 404 if the execution does not exist' do
+
+      exid = @exids.first
+p exid
+
+      r = @app.call(make_env(me: 'DELETE', path: "/executions/#{exid}"))
+
+      expect(r[0]).to eq(200)
+      expect(r[1]['Content-Type']).to eq('application/json')
+
+      j = JSON.parse(r[2].first)
+#pp j
+
+      expect(j['_links']['self']
+        ).to eq(
+          'href' => "/executions/#{exid}", 'method' => 'DELETE')
+
+      expect(j['exid']
+        ).to eq(exid)
+      expect(j['counts']
+        ).to eq(
+          'messages' => 0, 'executions' => 1, 'pointers' => 0,
+          'timers' => 1, 'traps' => 0)
+
+      expect(@app.unit.executions.map(:exid).sort
+        ).to eq(@exids - [ exid ])
+    end
+
+    it 'goes 200 if the execution exists'
+  end
 end