+++ spec/spec_helper.rb
@@ -15,7 +15,7 @@ module Helpers
 
   def make_env(opts)
 
-    me = opts[:method] || 'GET'
+    me = opts[:method] || opts[:me] || 'GET'
     ho = opts[:host] || '127.0.0.1:7006'
     pa = opts[:path] || opts[:p] || '/'
     qs = opts[:query] || opts[:qs] || ''
@@ -25,8 +25,7 @@ module Helpers
     body = JSON.dump(body) if body && ! body.is_a?(String)
     ri = body ? StringIO.new(body) : nil
 
-    {
-      'REQUEST_METHOD' => me,
+    { 'REQUEST_METHOD' => me,
       'PATH_INFO' => pa,
       'REQUEST_PATH' => pa,
       'QUERY_STRING' => qs,
@@ -35,8 +34,7 @@ module Helpers
       'HTTP_HOST' => ho,
       'HTTP_VERSION' => 'HTTP/1.1',
       'rack.url_scheme' => 'http',
-      'rack.input' => ri
-    }
+      'rack.input' => ri }
   end
 
   def jdump(o)